requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==22.1.0
aiohttp==3.7.4
aiormq==3.3.1
altgraph==0.17.4
annotated-types==0.7.0
APScheduler==3.7.0
asgiref==3.6.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
bidict==0.22.0
blis==1.2.0
boto3==1.26.39
botocore==1.29.39
CacheControl==0.12.11
cachetools==4.2.4
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cloudpathlib==0.20.0
cloudpickle==1.6.0
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
confection==0.1.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.11.0
cymem==2.0.7
dask==2022.2.0
distlib==0.3.6
Django==4.2.21
django-cors-headers==4.7.0
django-environ==0.9.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==1.16.0
docopt==0.6.2
dotenv==0.9.9
drf-yasg==1.20.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fbmessenger==6.0.0
filelock==3.9.0
fire==0.5.0
flatbuffers==2.0.7
fsspec==2025.2.0
future==0.18.2
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==2.0.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==20.1.0
h5py==3.7.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.4
importlib-metadata==5.2.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==4.4.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
mpmath==1.3.0
msgpack==1.0.4
multidict==5.2.0
murmurhash==1.0.9
netifaces==0.11.0
networkx==2.6.3
numpy==1.22.4
numpy-typing==1.1.1
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.4.3
partd==1.3.0
pathy==0.10.1
pefile==2023.2.7
Pillow==9.3.0
platformdirs==2.6.2
pluggy==1.0.0
preshed==3.0.8
prompt-toolkit==2.0.10
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.2
pyTelegramBotAPI==4.8.0
pytest==7.2.0
pytest-django==4.5.2
pytest-ordering==0.6
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==1.1.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
pywin32-ctypes==0.2.3
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.2.1
rasa-sdk==3.2.2
redis==3.5.3
regex==2022.4.24
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.25.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
safetensors==0.5.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
sentence-transformers==3.4.1
sentry-sdk==1.3.1
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==6.3.0
spacy==3.8.3
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.45
sqlparse==0.4.3
srsly==2.4.5
sympy==1.13.1
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminaltables==3.1.10
thinc==8.3.4
threadpoolctl==3.1.0
tokenizers==0.21.0
tomli==2.0.1
toolz==0.12.0
torch==2.6.0
tqdm==4.64.1
transformers==4.48.3
twilio==6.50.1
typeguard==2.13.3
typer==0.4.2
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==2.1
ujson==5.6.0
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.17.1
wasabi==0.10.1
wcwidth==0.2.5
weasel==0.4.1
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
xlrd==2.0.1
yarl==1.8.2
zipp==3.11.0
